+++ package.json
@@ -44,8 +44,9 @@
     "build": "npm run build:clean && npm run build:esm && npm run build:cjs && npm run build:universal-package && npm run build:chmod",
     "inspect": "npm run build && mcp-inspector -- dist/esm/index.js",
     "prettier": "prettier",
-    "check": "npm run build && npm run check:types && npm run check:lint && npm run check:format",
+    "check": "npm run build && npm run check:types && npm run check:lint && npm run check:format && npm run check:dependencies",
     "check:lint": "eslint .",
+    "check:dependencies": "knip --strict",
     "check:format": "prettier -c .",
     "check:types": "tsc --noEmit --project tsconfig.json",
     "fix": "npm run fix:lint && npm run reformat",
@@ -66,12 +67,11 @@
     "@mongodb-js/oidc-mock-provider": "^0.11.3",
     "@redocly/cli": "^2.0.8",
     "@types/express": "^5.0.3",
-    "@types/http-proxy": "^1.17.16",
-    "@types/node": "^24.3.0",
+    "@types/node": "^24.5.2",
     "@types/proper-lockfile": "^4.1.4",
     "@types/semver": "^7.7.0",
-    "@types/simple-oauth2": "^5.0.7",
     "@types/yargs-parser": "^21.0.3",
+    "@typescript-eslint/parser": "^8.44.0",
     "@vitest/coverage-v8": "^3.2.4",
     "ai": "^4.3.17",
     "duplexpair": "^1.0.2",
@@ -79,10 +79,10 @@
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-prettier": "^5.5.4",
     "globals": "^16.3.0",
+    "knip": "^5.63.1",
     "mongodb-runner": "^5.9.2",
     "ollama-ai-provider": "^1.2.0",
     "openapi-types": "^12.1.3",
-    "openapi-typescript": "^7.9.1",
     "prettier": "^3.6.2",
     "proper-lockfile": "^4.1.2",
     "semver": "^7.7.2",
@@ -92,20 +92,18 @@
     "typescript-eslint": "^8.41.0",
     "uuid": "^11.1.0",
     "vitest": "^3.2.4",
-    "yaml": "^2.8.1"
+    "@vitest/eslint-plugin": "^1.3.4",
+    "mongodb": "^6.19.0"
   },
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.17.4",
     "@mongodb-js/device-id": "^0.3.1",
-    "@mongodb-js/devtools-connect": "^3.9.3",
     "@mongodb-js/devtools-proxy-support": "^0.5.2",
     "@mongosh/arg-parser": "^3.14.0",
     "@mongosh/service-provider-node-driver": "~3.12.0",
-    "@vitest/eslint-plugin": "^1.3.4",
     "bson": "^6.10.4",
     "express": "^5.1.0",
     "lru-cache": "^11.1.0",
-    "mongodb": "^6.19.0",
     "mongodb-connection-string-url": "^3.0.2",
     "mongodb-log-writer": "^2.4.1",
     "mongodb-redact": "^1.2.0",
