+++ src/common/logger.ts
@@ -339,15 +339,3 @@ export class CompositeLogger extends LoggerBase {
         this.attributes[key] = value;
     }
 }
-
-export class NullLogger extends LoggerBase {
-    protected type?: LoggerType;
-
-    constructor() {
-        super(undefined);
-    }
-
-    protected logCore(): void {
-        // No-op logger, does not log anything
-    }
-}
