+++ tests/integration/common/connectionManager.test.ts
@@ -2,12 +2,13 @@ import type {
     ConnectionManagerEvents,
     ConnectionStateConnected,
     ConnectionStringAuthType,
-    TestConnectionManager,
 } from "../../../src/common/connectionManager.js";
+
 import { MCPConnectionManager } from "../../../src/common/connectionManager.js";
 import type { UserConfig } from "../../../src/common/config.js";
 import { describeWithMongoDB } from "../tools/mongodb/mongodbHelpers.js";
 import { describe, beforeEach, expect, it, vi, afterEach } from "vitest";
+import { type TestConnectionManager } from "../../utils/index.js";
 
 describeWithMongoDB("Connection Manager", (integration) => {
     function connectionManager(): TestConnectionManager {
