+++ tests/unit/common/session.test.ts
@@ -2,7 +2,8 @@ import type { Mocked } from "vitest";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { NodeDriverServiceProvider } from "@mongosh/service-provider-node-driver";
 import { Session } from "../../../src/common/session.js";
-import { config, driverOptions } from "../../../src/common/config.js";
+import { config } from "../../../src/common/config.js";
+import { driverOptions } from "../../integration/helpers.js";
 import { CompositeLogger } from "../../../src/common/logger.js";
 import { MCPConnectionManager } from "../../../src/common/connectionManager.js";
 import { ExportsManager } from "../../../src/common/exportsManager.js";
