+++ tests/unit/resources/common/debug.test.ts
@@ -2,7 +2,8 @@ import { beforeEach, describe, expect, it } from "vitest";
 import { DebugResource } from "../../../../src/resources/common/debug.js";
 import { Session } from "../../../../src/common/session.js";
 import { Telemetry } from "../../../../src/telemetry/telemetry.js";
-import { config, driverOptions } from "../../../../src/common/config.js";
+import { config } from "../../../../src/common/config.js";
+import { driverOptions } from "../../../integration/helpers.js";
 import { CompositeLogger } from "../../../../src/common/logger.js";
 import { MCPConnectionManager } from "../../../../src/common/connectionManager.js";
 import { ExportsManager } from "../../../../src/common/exportsManager.js";
